refactor(vm): split execution engine
